+++ docs/source/changelog.rst
@@ -2,6 +2,20 @@
 Changelog
 #########
 
+1.3.1 [Processor Linux SDK 6.1]
+===============================
+**Added**
+
+#. Added imagenet python example.
+
+**Changed**
+
+#. Trace data now also include dataQ/minValue/maxValue info.
+
+#. ssd_multibox example now can also run the whole netework on a single core.
+
+#. Shipped network binaries are in new file format, of size 484384 bytes.
+
 1.3.0 [Processor Linux SDK 5.3]
 ===============================
 **Added**
@@ -30,7 +44,7 @@ Changelog
     Enable a two phase approach to generating execution graphs. Use the
     following API function to enable timestamp generation:
 
-    .. code::
+    .. code:: cpp
 
         bool EnableTimeStamps(const std::string& file = "timestamp.log", size_t num_frames=32);
 
@@ -38,7 +52,7 @@ Changelog
 
 #. Added Python 3 bindings for TIDL API. See the ``examples/pybind`` directory for examples of using the Python bindings. Set PYTHONPATH to the location of ``tidl.so``.
 
-   .. code::
+   .. code:: bash
 
         root@am57xx-evm:~# export PYTHONPATH=/home/root/tidl-api/tidl_api
         root@am57xx-evm:~# python3